requirements.txt
Flask
networkx
matplotlib
Pillow
